feat: Add Horizontal scroll
This adds a new default behavior, horizontal scroll:
After activating scroll mode on a scrollable element,
scroll the trackball horizontally *with pressing Shift key* to scroll
the element horizontally (no vertical movement).

This behavior - 1) Requiring pressing Shift key, 2) moving in only 1
axis - is chosen as the behavior of scroll wheel.


While it can be made to be scrolled in both axes, it feels little
awkward and distracting to follow texts. I don't know if there's a use
case for it to be feel more natural to scroll both axis at the same
time (maybe like infinite canvas?), though it's easy to implement it
- just set `canScrollV = true` at all times.